+++ modules/experimental/_pkgs/ollama-cuda.nix
@@ -1,0 +1,7 @@
+{ inputs }:
+final: prev: {
+    ollama-cuda = (import inputs.nixpkgs-ollama {
+        inherit (final) system;
+        config.allowUnfree = true;
+    }).ollama-cuda;
+}